`[Configure/Queues]` Add ability to add IAM policies.
