+++ frontend/src/pages/Configure/HeadNode.js
@@ -17,7 +17,6 @@ import { findFirst, getIn } from '../../util'
 // UI Elements
 import {
   Box,
-  Button,
   Container,
   ExpandableSection,
   FormField,
@@ -32,7 +31,7 @@ import {
 import { setState, getState, useState, clearState, updateState, ssmPolicy } from '../../store'
 
 // Components
-import { ActionsEditor, InstanceSelect, RootVolume, SecurityGroups, SubnetSelect } from './Components'
+import { ActionsEditor, InstanceSelect, RootVolume, SecurityGroups, SubnetSelect, IamPoliciesEditor } from './Components'
 import HelpTooltip from '../../components/HelpTooltip'
 
 // Constants
@@ -246,42 +245,6 @@ function DcvSettings() {
   </div>
 }
 
-function IamPoliciesEditor() {
-  const policiesPath = [...headNodePath, 'Iam', 'AdditionalIamPolicies']
-  const policies = useState(policiesPath) || [];
-  const policyPath = ['app', 'wizard', 'headNode', 'iamPolicy'];
-  const policy = useState(policyPath) || '';
-
-  const addPolicy = () => {
-    updateState(policiesPath, (existing) => [...(existing || []), {Policy: policy}])
-    setState(policyPath, "");
-  }
-
-  const removePolicy = (index) => {
-    setState(policiesPath, [...policies.slice(0, index), ...policies.slice(index + 1)]);
-    if(policies.length === 0)
-      clearState(policiesPath)
-  }
-
-  return <SpaceBetween direction="vertical" size="s">
-    <FormField errorText={findFirst(policies, x => x.Policy === policy) ? "Policy already added." : ""}>
-      <SpaceBetween direction="horizontal" size="s">
-        <div style={{width: "400px"}}>
-          <Input
-            placeholder="arn:aws:iam::aws:policy/SecretsManager:ReadWrite"
-            value={policy}
-            onChange={({detail}) => setState(policyPath, detail.value)} />
-        </div>
-        <Button onClick={addPolicy} disabled={policy.length === 0 || findFirst(policies, x => x.Policy === policy)}>Add</Button>
-      </SpaceBetween>
-    </FormField>
-    {policies.map((p, i) => p.Policy !== ssmPolicy && <SpaceBetween key={p.Policy} direction="horizontal" size="s">
-      <div style={{width: "400px"}}>{p.Policy}</div>
-      <Button onClick={() => removePolicy(i)}>Remove</Button>
-    </SpaceBetween>)}
-  </SpaceBetween>
-}
-
 function HeadNode() {
   const imdsSecuredPath = [...headNodePath, 'Imds', 'Secured'];
   const imdsSecured = useState(imdsSecuredPath);
@@ -341,9 +304,8 @@ function HeadNode() {
       <ExpandableSection header="Advanced options">
         <ActionsEditor basePath={headNodePath} errorsPath={errorsPath}/>
         <ExpandableSection header="IAM Policies">
-          <IamPoliciesEditor />
+          <IamPoliciesEditor basePath={headNodePath} />
         </ExpandableSection>
-
       </ExpandableSection>
     </SpaceBetween>
   </Container>
